+++ src/cc/arch/aarch64/aarch64.h
@@ -303,4 +303,4 @@
 #define FCVTZS(o1, o2)  EMIT_ASM2("fcvtzs", o1, o2)  // int <- float
 #endif
 
-void mov_immediate(const char *dst, intptr_t value, bool b64);
+void mov_immediate(const char *dst, int64_t value, bool b64);
